GlusterFS role fixes (#174)
* Rename storage-cns to storage-glusterfs

"CNS" is a product, not a project or technology. To be accurate in
all supported configurations, we should say we are integrating
GlusterFS storage.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>

* Rename storage-glusterfs.yml template

This is not strictly a YAML file, it is a Jinja2 template to generate a
YAML file. Rename accordingly.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>

* Add heketi_url to storage-glusterfs role

Relying on heketi_route causes this role to not work outside of
OpenShift. This allows Kubernetes users to specify a heketi URL. If a
heketi_route_name is specified, the generated heketi_route takes
precedence over heketi_url.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>