GlusterFS: increase default snapfactor to prevent cloning issues
By default the 'snapfactor' is set to 1.0, which means that there is
very little space reserved for modifications to cloned volumes. By
increasing the 'snapfactor' to 3.0, the requested PVC will be backed by
a larger (LVM) thin-pool on the Gluster storage servers and it is less
likely that the thin-pool will run out of space when a few PVCs are
cloned from the original.

The 'snapfactor' option is only available in the external provisioner,
and was introduced with kubernetes-incubator/external-storage#648.

Signed-off-by: Niels de Vos <ndevos@redhat.com>