Render a jinja2 template for the kubevirt manifest (#279)
This change allows kubevirt-ansible to consume a jinja2 template rather than a pre-rendered manifest from kubevirt.

For now, we include a jinja2 version, but the aim is to have upstream kubevirt maintain it. Otherwise we'll have to set up some system to automatically create a new j2 when upstream changes.