Variable to specify upstream Kubernetes version
We need a way to specify the version of Kuberbetes that is installed
because KubeVirt may not work on the latest upstream version.
For example, upstream is now on 1.11, but there are incompatibility
problems with latest, so we need the ability to install the previous
stable version, 1.10, until issues are resolved.