Use spc_t type for SELinux context of sriovdp container
This container needs access to /var/lib/kubelet/device-plugins/ to
create a Unix socket that kubelet then uses to talk to the plugin.

Ideally, we would have a special type crafted for device plugins that
would limit access to just this directory. In the meantime, we are
going to use spc_t which stands for Super Privileged Container.

More info on spc_t here: https://danwalsh.livejournal.com/74754.html