metrics: add role and playbook
Add role and playbook to deploy the kubevirt metrics collector.
Until we have proper integration with OCP/OKD, we use a hack
to sneak in the existing monitoring, pretending to be
a node_exporter.

For this reason the deploy must be enabled explicitely using
the variable `kubevirt_metrics_collector_enabled`.

Signed-off-by: Francesco Romani <fromani@redhat.com>